+++ hslua/src/HsLua/Types.hs
@@ -9,9 +9,13 @@ Portability : non-portable (depends on GHC)
 Types for working with Lua.
 -}
 module HsLua.Types
-  ( module HsLua.Types.Peekable
+  ( module HsLua.Types.Exposable
+  , module HsLua.Types.Invokable
+  , module HsLua.Types.Peekable
   , module HsLua.Types.Pushable
   ) where
 
+import HsLua.Types.Exposable
+import HsLua.Types.Invokable
 import HsLua.Types.Peekable
 import HsLua.Types.Pushable